+++ src/main/java/server/MainHandler.java
@@ -1,17 +1,19 @@
-package Server;
+package server;
 
-import Common.Command;
-import Common.FileMessage;
-import Common.FileRequest;
+import common.Command;
+import common.FileMessage;
+import common.FileRequest;
 import io.netty.channel.ChannelHandlerContext;
 import io.netty.channel.ChannelInboundHandlerAdapter;
 import io.netty.util.ReferenceCountUtil;
 
 import java.io.File;
 import java.nio.file.Files;
+import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.nio.file.StandardOpenOption;
 import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.List;
 
 public class MainHandler extends ChannelInboundHandlerAdapter {
@@ -36,7 +38,22 @@ public class MainHandler extends ChannelInboundHandlerAdapter {
             }
             if (msg instanceof FileMessage) {
                 FileMessage fm = (FileMessage) msg;
-                Files.write(Paths.get(storage + fm.getFilename()), fm.getData(), StandardOpenOption.CREATE);
+                Path path = Paths.get(storage + fm.getFilename());
+                if (!Files.exists(path)) {
+                    Files.write(Paths.get(storage + fm.getFilename()), fm.getData(), StandardOpenOption.CREATE);
+                } else{
+                    int count = 1;
+                    while (true){
+                        String[] strings = fm.getFilename().split("");
+                        String title = nameBuilder(strings, count);
+                        Path name = Paths.get(storage + title);
+                        if (Files.exists(name)) count++;
+                        else {
+                            Files.write(Paths.get(storage + title), fm.getData(), StandardOpenOption.CREATE);
+                            break;
+                        }
+                    }
+                }
                 update();
                 Command cmd = new Command("update", results);
                 ctx.writeAndFlush(cmd);
@@ -64,12 +81,6 @@ public class MainHandler extends ChannelInboundHandlerAdapter {
         }
     }
 
-    @Override
-    public void exceptionCaught(ChannelHandlerContext ctx, Throwable cause) throws Exception {
-        cause.printStackTrace();
-        ctx.close();
-    }
-
     private void update(){
         results.clear();
         File[] files = new File(storage).listFiles();
@@ -79,5 +90,21 @@ public class MainHandler extends ChannelInboundHandlerAdapter {
                 results.add(file.getName());
             }
         }
+    }
+
+    private String nameBuilder(String[] strings, int count){
+        ArrayList<String> arrayList = new ArrayList<>(Arrays.asList(strings));
+        arrayList.add(arrayList.size()-4, "(" + count + ")");
+        StringBuilder builder = new StringBuilder(arrayList.size());
+        for(String s: arrayList) {
+            builder.append(s);
+        }
+        return builder.toString();
+    }
+
+    @Override
+    public void exceptionCaught(ChannelHandlerContext ctx, Throwable cause) throws Exception {
+        cause.printStackTrace();
+        ctx.close();
     }
 }
